\name{dssatcalibrator-api}
\alias{dssatcalibrator-api}
\alias{\%||\%}
\alias{acquire_soil_profile}
\alias{acquire_wth}
\alias{active_parameters}
\alias{ADAPTER_REGISTRY}
\alias{all_parameters}
\alias{anchor_correction}
\alias{anova_variance_share}
\alias{apply_delta_from_analog}
\alias{apply_hierarchical_priors}
\alias{apply_identifiability_gate}
\alias{apply_sparse_config}
\alias{apply_trait_priors}
\alias{assimilate}
\alias{BATCH_FILE}
\alias{build_residuals}
\alias{build_sources}
\alias{build_staged_configs}
\alias{calibrate}
\alias{calibrate_quick}
\alias{calibrate_staged}
\alias{combine_runs}
\alias{combined_mode}
\alias{coverage}
\alias{crop_for}
\alias{cultivar_field_map}
\alias{default_stages}
\alias{ecotype_field_map}
\alias{edit_cultivar}
\alias{edit_ecotype}
\alias{edit_filex}
\alias{edit_soil}
\alias{edit_species}
\alias{edit_weather}
\alias{enkf_assimilate}
\alias{ensemble_percentiles}
\alias{evaluate_design}
\alias{evaluate_thetas}
\alias{expand_parameter_specs}
\alias{extract_soil_profile}
\alias{fixed_parameters}
\alias{forcing_apply}
\alias{forecast_lai}
\alias{fuser_collect}
\alias{fuser_resolve_conflicts}
\alias{GENETIC_GROUPS}
\alias{has_informative_prior}
\alias{identifiability}
\alias{influential_params}
\alias{lead_time_table}
\alias{load_config}
\alias{log_prior_one}
\alias{log_prior_vec}
\alias{make_quick_dirty_config}
\alias{make_report}
\alias{MASS_ginv}
\alias{metrics}
\alias{MISSING}
\alias{normalize_treatments}
\alias{nowcast}
\alias{obs_experiments}
\alias{obs_variables}
\alias{observation_fuser}
\alias{observations}
\alias{observations_from_csv}
\alias{observations_from_dssat}
\alias{observations_from_sources}
\alias{parameter_space_from_config}
\alias{parse_cultivars}
\alias{parse_evaluate}
\alias{parse_fields}
\alias{parse_header_boundaries}
\alias{parse_plantgro}
\alias{parse_cv_config}
\alias{parse_summary}
\alias{parse_treatments}
\alias{planting_dates}
\alias{posterior_summary}
\alias{print.impact_atlas_result}
\alias{ps_clip}
\alias{ps_ndim}
\alias{ps_to_theta}
\alias{read_csv}
\alias{read_cul_calibration_bounds}
\alias{read_cultivar_values}
\alias{read_ecotype_values}
\alias{read_filea}
\alias{read_filet}
\alias{recalibrate}
\alias{recommend_observations}
\alias{resolve_cores}
\alias{resolve_dssat_paths}
\alias{resolve_dssat_root}
\alias{resolve_exe}
\alias{resolve_template_dir}
\alias{run_abc_smc}
\alias{run_bayesopt}
\alias{run_cross_validation}
\alias{run_dream}
\alias{run_es_mda}
\alias{run_glue}
\alias{run_history_matching}
\alias{run_impact_atlas}
\alias{run_many}
\alias{run_mcmc}
\alias{run_morris}
\alias{run_nsga2}
\alias{run_optimizer}
\alias{run_sensitivity}
\alias{run_smc_pf}
\alias{run_sobol}
\alias{run_surrogate}
\alias{sample_design}
\alias{sample_one}
\alias{sample_prior_design}
\alias{scaffold_crop}
\alias{SCHEMA}
\alias{SCHEMA_EXTENDED}
\alias{score}
\alias{sd_pop}
\alias{source_canopy_temperature}
\alias{source_farm_management}
\alias{source_farm_phenology}
\alias{source_field_measurements}
\alias{source_modis_lai}
\alias{source_sentinel2_lai}
\alias{source_soil_moisture_iot}
\alias{source_uav_multispectral}
\alias{spawn_and_run}
\alias{spawn_result}
\alias{spawn_results_for}
\alias{src_error_model}
\alias{src_error_model.canopy_temperature}
\alias{src_error_model.farm_management}
\alias{src_error_model.farm_phenology}
\alias{src_error_model.field_measurements}
\alias{src_error_model.modis_lai}
\alias{src_error_model.sentinel2_lai}
\alias{src_error_model.soil_moisture_iot}
\alias{src_error_model.uav_multispectral}
\alias{src_fetch}
\alias{src_fetch.canopy_temperature}
\alias{src_fetch.farm_management}
\alias{src_fetch.farm_phenology}
\alias{src_fetch.field_measurements}
\alias{src_fetch.modis_lai}
\alias{src_fetch.sentinel2_lai}
\alias{src_fetch.soil_moisture_iot}
\alias{src_fetch.uav_multispectral}
\alias{src_quality_filter}
\alias{src_quality_filter.default}
\alias{src_quality_filter.sentinel2_lai}
\alias{src_variable_mapping}
\alias{src_variable_mapping.canopy_temperature}
\alias{src_variable_mapping.farm_management}
\alias{src_variable_mapping.farm_phenology}
\alias{src_variable_mapping.field_measurements}
\alias{src_variable_mapping.modis_lai}
\alias{src_variable_mapping.sentinel2_lai}
\alias{src_variable_mapping.soil_moisture_iot}
\alias{src_variable_mapping.uav_multispectral}
\alias{stepwise_select}
\alias{structural_adequacy}
\alias{summary_fit_table}
\alias{systematic_resample}
\alias{theta_hash}
\alias{unmatched_variables}
\alias{validate_config}
\alias{validate_cv}
\alias{validate_loeo}
\alias{variable_maps}
\alias{write_dssbatch}
\alias{write_cv_report}
\alias{yyddd_to_date}
\title{DSSAT Calibration API}
\description{
Public configuration, parameter-space, calibration-engine, observation,
assimilation, spawning, validation, and reporting entry points. The README
documents end-to-end workflows and the R/Python parity contract.
}
